Add `umd` & `es` targets
