+++ .babelrc
@@ -1,3 +1,25 @@
 {
-  "presets": ["env"]
+  "env":{
+    "es": {
+      "plugins": [
+        "babel-plugin-annotate-pure-calls"
+      ]
+    },
+    "umd": {
+      "presets": [
+        ["babel-preset-env",{
+          "modules": false,
+          "targets":{
+            "chrome": 43,
+            "firefox": 43,
+            "safari": 10,
+            "ios": 10,
+            "opera": 40,
+            "edge": 13,
+            "ie": 9
+          }
+        }]
+      ]
+    }
+  }
 }
