+++ .gitignore
@@ -26,3 +26,4 @@ yarn.lock
 .Trashes
 
 #dist/
+/es
