+++ README.md
@@ -1,2 +1,17 @@
 # wc-link
 `<wc-link>` web component
+
+## How to
+
+### Demo
+```bash
+npm start
+```
+
+### Use
+```html
+<script>window.customElements.define('wc-link', WCLink.Link)</script>
+<body>
+  <wc-link></wc-link>
+</body>
+```
